requirements.txt
aiohttp
discord.py<=1.2.5
async-sqlalchemy